--- a/.golangci.yml
+++ b/.golangci.yml
@@ -16,6 +16,7 @@ linters-settings:
           - "!$test"
         allow:
           - $gostd
+          - github.com/ugorji/go/codec
       Test:
         files:
           - $test