--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -2,7 +2,10 @@ module github.com/fixme_my_friend/hw10_program_optimization
 
 go 1.22
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/stretchr/testify v1.7.0
+	github.com/ugorji/go/codec v1.2.12
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect