## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов


### Результаты benchstat

```
goos: darwin
goarch: arm64
pkg: github.com/romangricuk/otus-go/hw10_program_optimization
                    │      ./old      │                ./new                │
                    │     sec/op      │   sec/op     vs base                │
GetDomainStatCom-10   76098.409µ ± 0%   7.060µ ± 4%  -99.99% (p=0.000 n=10)
GetDomainStatGov-10   75985.769µ ± 1%   6.886µ ± 0%  -99.99% (p=0.000 n=10)
geomean                   76.04m        6.972µ       -99.99%

                    │       ./old       │                ./new                 │
                    │       B/op        │     B/op      vs base                │
GetDomainStatCom-10   132856.915Ki ± 0%   8.297Ki ± 0%  -99.99% (p=0.000 n=10)
GetDomainStatGov-10   132856.729Ki ± 0%   8.297Ki ± 0%  -99.99% (p=0.000 n=10)
geomean                    129.7Mi        8.297Ki       -99.99%

                    │      ./old      │                ./new                │
                    │    allocs/op    │ allocs/op   vs base                 │
GetDomainStatCom-10   1700090.00 ± 0%   57.00 ± 0%  -100.00% (p=0.000 n=10)
GetDomainStatGov-10   1700082.00 ± 0%   57.00 ± 0%  -100.00% (p=0.000 n=10)
geomean                   1.700M        57.00       -100.00%

```